+++ logya/generate.py
@@ -8,21 +8,21 @@ from logya.writer import DocWriter
 
 
 class Generate(Logya):
-    """Generate a Web site to deploy from current directory as source."""
+    """Generate a Web site to public from current directory as source."""
 
     def __init__(self, **kwargs):
         super(Generate, self).__init__(**kwargs)
         self.init_env()
-        self.writer = DocWriter(self.dir_deploy, self.template)
+        self.writer = DocWriter(self.dir_public, self.template)
 
         if not kwargs.get('keep'):
-            self.info('Remove existing deploy directory')
-            shutil.rmtree(self.dir_deploy, True)
+            self.info('Remove existing public directory')
+            shutil.rmtree(self.dir_public, True)
 
-        self.info('Generate site in directory: {}'.format(self.dir_deploy))
+        self.info('Generate site in directory: {}'.format(self.dir_public))
         if os.path.exists(self.dir_static):
             self.info('Copy static files')
-            copytree(self.dir_static, self.dir_deploy)
+            copytree(self.dir_static, self.dir_public)
 
         self.build()
         self.write()
@@ -36,11 +36,11 @@ class Generate(Logya):
         for doc in self.docs.values():
             self.writer.write(doc, self.get_doc_template(doc))
         self.info(
-            'Written {:d} documents to deploy directory'
+            'Written {:d} documents to public directory'
             .format(len(self.docs)))
 
         self.info('Write index files')
         self.write_index_files()
         self.info(
-            'Written {:d} index files to deploy directory'
+            'Written {:d} index files to public directory'
             .format(len(self.index)))